fix backend dist overide funsor dist
