+++ test/conftest.py
@@ -4,6 +4,7 @@
 import numpy as np
 
 import funsor.util
+funsor.util.set_backend("torch")
 
 
 def _disallow_set_backend(*args):
